requirements.txt
torch
torchvision
facenet-pytorch
numpy
lmdb
faiss-cpu
pillow
aiogram
opencv-python-headless
kagglehub
scipy
pandas
